Update datasets.py - numpy int update
The error you're encountering is caused by the use of np.int, which is deprecated and no longer available in recent versions of NumPy. The solution is to replace np.int with the built-in int type or a specific NumPy integer type like np.int64 or np.int32.

he error occurs in the file /kaggle/working/edgeai-yolov5/utils/datasets.py, line 428, where np.int is being used. You will need to modify this line to use the standard Python int instead of np.int.